+++ main.py
@@ -1,0 +1,4 @@
+from appBase import Application
+
+app = Application()
+app.mainloop()
